+++ x-pack/plugins/endpoint/public/applications/endpoint/alerts/view/url_from_query_params.ts
@@ -5,7 +5,8 @@
  */
 
 import querystring from 'querystring';
-import { AlertingIndexUIQueryParams, EndpointAppLocation } from '../../types';
+import { AlertingIndexUIQueryParams } from '../../../../../common/alerts/types';
+import { EndpointAppLocation } from '../../types';
 
 /**
  * Return a relative URL for `AlertingIndexUIQueryParams`.
