+++ x-pack/plugins/endpoint/server/alerts/handlers/details/lib/pagination.ts
@@ -5,12 +5,18 @@
  */
 
 import { GetResponse, SearchResponse } from 'elasticsearch';
-import { AlertEvent, AlertHits, AlertAPIOrdering } from '../../../../../common/types';
+import {
+  AlertEvent,
+  AlertHits,
+  AlertAPIOrdering,
+  AlertSearchQuery,
+  SearchCursor,
+  AlertDetailsRequestParams,
+} from '../../../../../common/alerts/types';
 import { AlertConstants } from '../../../../../common/alert_constants';
 import { EndpointConfigType } from '../../../../config';
 import { searchESForAlerts, Pagination } from '../../lib';
-import { AlertSearchQuery, SearchCursor, AlertDetailsRequestParams } from '../../types';
-import { BASE_ALERTS_ROUTE } from '../..';
+import { BASE_ALERTS_ROUTE } from '../../routes';
 import { RequestHandlerContext } from '../../../../../../../../src/core/server';
 import { Filter } from '../../../../../../../../src/plugins/data/server';
