+++ x-pack/plugins/endpoint/server/alerts/handlers/details/schemas.ts
