--- x-pack/plugins/endpoint/server/routes/alerts/details/index.ts
@@ -1,8 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
- */
-
-export { alertDetailsReqSchema } from './schemas';
-export { alertDetailsHandlerWrapper } from './handlers';
